% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantCare_app.R
\name{GUI}
\alias{GUI}
\title{Shiny app that tells you when to water your plant}
\usage{
GUI()
}
\arguments{
\item{...}{No parameters yet}
}
\value{
Pop-up with the user interface
}
\description{
An interactive app that gives advice on whether to water, fertilize, or change light conditions for common houseplants. Assumes the user is close to their plant (i.e., to feel whether the soil is wet and see if the leaves are droopy).
}
